@@ -4,6 +4,7 @@ from sqlglot import alias, exp
 from sqlglot.optimizer.qualify_columns import Resolver
 from sqlglot.optimizer.scope import Scope, traverse_scope
 from sqlglot.schema import ensure_schema
+from sqlglot.errors import OptimizeError
 
 # Sentinel value that means an outer query selecting ALL columns
 SELECT_ALL = object()
@@ -48,21 +49,31 @@ def pushdown_projections(expression, schema=None, remove_unused_selections=True)
             parent_selections = {SELECT_ALL}
 
         if isinstance(scope.expression, exp.SetOperation):
-            left, right = scope.union_scopes
-            referenced_columns[left] = parent_selections
+            set_op = scope.expression
+            if not (set_op.kind or set_op.side):
+                # Do not optimize this set operation if it's using the BigQuery specific
+                # kind / side syntax (e.g INNER UNION ALL BY NAME) which changes the semantics of the operation
+                left, right = scope.union_scopes
+                if len(left.expression.selects) != len(right.expression.selects):
+                    scope_sql = scope.expression.sql()
+                    raise OptimizeError(
+                        f"Invalid set operation due to column mismatch: {scope_sql}."
+                    )
 
-            if any(select.is_star for select in right.expression.selects):
-                referenced_columns[right] = parent_selections
-            elif not any(select.is_star for select in left.expression.selects):
-                if scope.expression.args.get("by_name"):
-                    referenced_columns[right] = referenced_columns[left]
-                else:
-                    referenced_columns[right] = [
-                        right.expression.selects[i].alias_or_name
-                        for i, select in enumerate(left.expression.selects)
-                        if SELECT_ALL in parent_selections
-                        or select.alias_or_name in parent_selections
-                    ]
+                referenced_columns[left] = parent_selections
+
+                if any(select.is_star for select in right.expression.selects):
+                    referenced_columns[right] = parent_selections
+                elif not any(select.is_star for select in left.expression.selects):
+                    if scope.expression.args.get("by_name"):
+                        referenced_columns[right] = referenced_columns[left]
+                    else:
+                        referenced_columns[right] = [
+                            right.expression.selects[i].alias_or_name
+                            for i, select in enumerate(left.expression.selects)
+                            if SELECT_ALL in parent_selections
+                            or select.alias_or_name in parent_selections
+                        ]
 
         if isinstance(scope.expression, exp.Select):
             if remove_unused_selections:
