@@ -1,6 +1,6 @@
-import sys
 import datetime
 import math
+import sys
 import unittest
 
 from sqlglot import ParseError, alias, exp, parse_one
@@ -8,6 +8,19 @@ from sqlglot import ParseError, alias, exp, parse_one
 
 class TestExpressions(unittest.TestCase):
     maxDiff = None
+
+    def test_to_s(self):
+        self.assertEqual(repr(parse_one("5")), "Literal(this=5, is_string=False)")
+        self.assertEqual(repr(parse_one("5.3")), "Literal(this=5.3, is_string=False)")
+        self.assertEqual(repr(parse_one("True")), "Boolean(this=True)")
+        self.assertEqual(repr(parse_one("'  x'")), "Literal(this='  x', is_string=True)")
+        self.assertEqual(repr(parse_one("' \n  x'")), "Literal(this=' \\n  x', is_string=True)")
+        self.assertEqual(
+            repr(parse_one("   x ")), "Column(\n  this=Identifier(this=x, quoted=False))"
+        )
+        self.assertEqual(
+            repr(parse_one('"   x "')), "Column(\n  this=Identifier(this='   x ', quoted=True))"
+        )
 
     def test_arg_key(self):
         self.assertEqual(parse_one("sum(1)").find(exp.Literal).arg_key, "this")
@@ -55,7 +68,6 @@ class TestExpressions(unittest.TestCase):
             parse_one("ROW() OVER(Partition by y)"),
             parse_one("ROW() OVER (partition BY y)"),
         )
-        self.assertEqual(parse_one("TO_DATE(x)", read="hive"), parse_one("ts_or_ds_to_date(x)"))
         self.assertEqual(exp.Table(pivots=[]), exp.Table())
         self.assertNotEqual(exp.Table(pivots=[None]), exp.Table())
         self.assertEqual(
@@ -275,6 +287,16 @@ class TestExpressions(unittest.TestCase):
                 {
                     "`a-b`.`c`": parse_one("select 1"),
                 },
+                dialect="spark",
+            ).sql(),
+            "SELECT * FROM (SELECT 1) AS a /* source: a-b.c */",
+        )
+
+    def test_expand_with_lazy_source_provider(self):
+        self.assertEqual(
+            exp.expand(
+                parse_one('select * from "a-b"."C" AS a'),
+                {"`a-b`.c": lambda: parse_one("select 1", dialect="spark")},
                 dialect="spark",
             ).sql(),
             "SELECT * FROM (SELECT 1) AS a /* source: a-b.c */",
@@ -637,6 +659,12 @@ class TestExpressions(unittest.TestCase):
         self.assertTrue(all(isinstance(e, exp.Expression) for e in expression.walk()))
         self.assertTrue(all(isinstance(e, exp.Expression) for e in expression.walk(bfs=False)))
 
+    def test_str_position_order(self):
+        str_position_exp = parse_one("STR_POSITION('mytest', 'test')")
+        self.assertIsInstance(str_position_exp, exp.StrPosition)
+        self.assertEqual(str_position_exp.args.get("this").this, "mytest")
+        self.assertEqual(str_position_exp.args.get("substr").this, "test")
+
     def test_functions(self):
         self.assertIsInstance(parse_one("x LIKE ANY (y)"), exp.Like)
         self.assertIsInstance(parse_one("x ILIKE ANY (y)"), exp.ILike)
@@ -841,6 +869,7 @@ class TestExpressions(unittest.TestCase):
 
     def test_convert(self):
         from collections import namedtuple
+
         import pytz
 
         PointTuple = namedtuple("Point", ["x", "y"])
